Update README.md
This project is the 4th Assignment for "Applied ML in Python" course from Uni of Michigan (Coursera). The main idea is to find the best ML approach for predicting probabilities of that the corresponding blight ticket will be paid on time. The Limitations of the task were the time - the whole code has to be run within 10 min + WOE (Weight Of Evidence Encoder) didn't work on Jupiter. So, in the submitted version I used CatBoostEncoder() and I ran LogReg, which in sum brought me right to the edge of submissions (AUC 0.73). Whereas, here I want to present the initial code with WOE encoder and XGBClassifier()/GradientBoostingClassifier() models which takes more time to run (more than 10 min), but bring more accurate outputs (AUC 0.93).